+++ CHANGELOG.md
@@ -6,6 +6,7 @@ See [Upgrading] for details how to upgrade.
 
 - Move inline documentation to docs/help and convert to markdown, #921
 - Move imap related methods to ImapConnection class, #937
+- Fix assignment field and quick search dropdowns, #943, #916, #942
 
 [3.9.6]: https://github.com/eventum/eventum/compare/v3.9.5...master